Manage IAM user for metrics tool
We want to monitor the certificate validity on our CloudFront
Distributions.  This creates an IAM user and adds it to a group which is
only allowed to list CloudFront distributions.

We have copied the pattern for AWS key rotation that was used for the
SES SMTP user. Ideally, this app could be extracted to a Bosh release
and an instance profile could be used, eliminating the need for keys
altogether.