Add CDN TLS gauge to metrics tool
We add this new metric gauge because we want to alert when a certificate
used by a CDN instance is due to expire soon. The aim is very similar to
the existing TLS gauge, which checks the validity of our public-facing
platform certificates. Because of this we have split the certificate
checking parts of the code and made it an interface, so we can more
easily unit test the original TLS gauge and our new one.

The CDN TLS gauge needs to list CloudFront distributions, so we provide
AWS access keys to the app. Ideally, we could run this as a Bosh release
and use an instance profile, therefore eliminating the need to manage
keys.

We collect pairs of domains consisting of the CloudFront domain and the
DNS Alias/CNAME because we can't rely on the Alias/CNAME being mapped to
CloudFront at the time of the check.

Fake CertChecker generated with `go generate ./...`