Add Datadog monitor for CDN certificate expiration
We want to know when certificates for tenant's cusom domains are due to
expire (or if they are invalid). This monitor makes use of the new
`cdn.tls.certificates.validity` metric, and is based off the existing
TLS monitor in the same file.

We are only emitting the metric hourly, so we alert on metrics collected
over a 4 hour period and allow 4 hours before a `NO_DATA` alert is
generated.