Use custom loggregator BOSH release
We added a drain script to the loggregator_trafficcontroller job to make sure
traffic controller is stopped before the consul-agent. Previously when the VM
was stopped we got CF-StatsUnavailable from the CF API which was caused by the
following traffic controller errors:

```
2018/01/23 15:13:28 Could not get app information: [Get
https://cloud-controller-ng.service.cf.internal:9023/internal/v4/log_access/6ba02750-f073-444f-ba73-ee3cf4a02ec6:
dial tcp: lookup cloud-controller-ng.service.cf.internal on 10.0.0.2:53: no such host
```

We saw these errors when the consul-agent was stopped before the
trafficcontroller service and stopped providing DNS resolution.