Implement logic to rotate CF internal certs
We want to be able to rotate the CF internal certificates in a
automated fashion, in a similar way we rotate other CF credentials.

Some public facing certs, like app or sys domains, won't be rotated
with this operation.

For this we extend the generate-cf-certs.sh script, to support
rotation. An additional argument is passed: create|rotate.

when:
   - the operation is "rotate"
   - the certificate exists
   - and the certificate is not blacklisted (app|sys)
it will rename the existing certificate to add a `_old` prefix and
generate a new one.

These files `<cn>_old.*` would be later available for terraform
and spruce with the `_old` prefix, so we can implement some
certificate transition process that requires both certificates
(e.g. the JWT key).

We add a pair of jobs in the certificates group:
 - rotate-cf-internal-certs: to run the rotation of certs
 - delete-old-cf-internal-certs: deletes the old certificates after
   the rotation, to be used after the new certs have been deployed.