Configure UAA JWT signing key with previous cert
When rotating certificates for UAA signing, we want to be able
to have a grace period to not make invalid every token generated
with the old certificate.

We can do this by setting multiple valid keys for key verification,
including the new and the old certificate, but using the new
as default for generation[1]

From the previous commit, after rotation of the certs, we might
have the old certificate in `secrets.uaa_jwt_signing_key_old`.
We pull that certificate as `previous` certificate.

The key `secrets.uaa_jwt_signing_key_old` might not exist and
be empty because the certificates have never been rotated or
the old certificates deleted. In that case we use
`secrets,uaa_jwt_signing_key` instead, that effectively
to expire immediatelly all duplicates it in a redundant way. That is
the simplest way of implementing conditionally add 1 or 2 certificates
with spruce.

[1] https://github.com/cloudfoundry/uaa-release/blob/develop/jobs/uaa/spec#L542