Do not require full window of CDN TLS metrics
We have been getting 'no data' alerts in prod for this monitor.  These
metrics are only collected every hour, which seems to be too sparse for
Datadog to consider it a 'full window' of data[1]. Also, Datadog advise
the 'no data' period be at least twice the window being evaluated.

[1] https://help.datadoghq.com/hc/en-us/articles/115002630843?input_string=no_data+alerts