Rename paas-usage-events-collector to paas-billing
The old name was difficult to remember and didn't reflect the fact that
this app now has an API to reporting billing.

This commit contains a `cf rename` command with a `FIXME` comment that
can be removed after it has been deployed in prod.

I haven't attempted to redirect the old hostname because we are the only
users and we can communicate the change on our team mailing list.