Run the smoke tests from paas-cf-smoke-tests (branch: fix_app_scale_tests_154477526)
We had to fix the smoke tests because of a tightly timed test which used to fail randomly.

Upstream PR: https://github.com/cloudfoundry/cf-smoke-tests/pull/39

In the "can be pushed, scaled and deleted" tests the application is upscaled to two instances. The tests wait until both application instances are in a running state and run up to 30 requests to check whether both instances get traffic.

The timing of these tests are really tight, as the route registration can be still pending even if the API returns a running state for the second instance and the requests are too quick (we are seeing only ~400ms for all 30 requests). In rare cases the tests fail although the route is eventually registered a little bit later (some hundreds of milliseconds).

We add some sleep between the requests, which adds an extra 5 seconds window for the new route registration to finish.