Fix wrong name for old jwt signing key
We made a typo in the name of the key when pulling the old jwt
signing key for UAA during the certificate rotation.

The name should be <cert-name> + _old + _key, as per [1].

The issue would be shadowed by spruce, as we pull a default value
if the key is not defined. We added a test for this specific
case in the manifest.

[1] https://github.com/alphagov/paas-cf/blob/0567d631f09a8b352fc8629375a81731f523069a/concourse/pipelines/create-cloudfoundry.yml#L1304-L1313