Remove old paas-usage-events-collector UAA client
The client was renamed in the preceding commit. This will result in UAA
creating a new client with the same credentials, so we need to manually
clean up the old client name.

The secret will remain in `cf-secrets.yml` but that doesn't matter so
much and is more fiddly to clean up.