Pull smoke-tests and CATs from our own branch
We pull the cf-smoke-tests and cf-acceptance-tests dependencies
directly from their repositories, instead of using the cf-release
submodules.

But we must pin the code to the "blessed version" for the right
combination of bosh releases.

In the cf-deployment[1] they pin it by using a semantic branch
in the cf-acceptance-tests. We will do a similar approach, by
forking the cf-acceptance-tests and the cf-smoke-tests, and
creating a branch based on our blessed version of cf-release
or cf-acceptance-tests.

For example, in our current case our deployment is based on
cf-release v283, so we create the branch gds-cf-release-v283.

Note: Our smoke-tests branch would also include the patch cloudfoundry/cf-smoke-tests#39

[1] https://github.com/cloudfoundry/cf-deployment/blame/9caa4c39962817d88665856753b2869eb5e1ae47/README.md#L238

Signed-off-by: Andrew Bruce <me@andrewbruce.net>