Update Redis plan
We want to enable backups of Elasticache Redis instances.

We have removed `automatic_failover_enabled` as a configurable property
as it will be hardcoded to `true` in the broker from now on. This is
because Elasticache does not allow backups for `t1.micro` and `t2.*`
instances if cluster mode is disabled; which means we have to enable
cluster mode (sharding); which means automatic failover has to be
enabled always.