Use ACM certs for apps and system domains
We want to use the preexisting ACM certificates for the apps and
system domains, that have been created in paas-terraform.

To do so, we use the `aws_acm_certificate`[1] data source from
terraform.

The outputs for the ARN are not being used, so we remove them.

Based on the code from #142501669