Temporary task to delete old app&sys certs
Note: This commit is meant to be reverted as soon the task is
executed in all the environments.

We want to delete the old app and system certs.

For that, we add a set of tasks that:
 - Will delete the certs from the cf-certs.tgz in the state-bucket
   if they exist.
 - Will run "terraform destroy" for the cf-certs terraform module,
 - Will delete the cf-certs.tfstate after.

In this commit we add again the minimum terraform files for
terraform/cf-certs required for terraform destroy to run.

We want to delete the cf-certs.tfstate and remove the dependency
with that resource, and not block new deployments. For that,
we use a task to download the S3 object if it exists, as the
S3 resource would block the pipeline if it does not exists.