Remove CF_SKIP_SSL_VALIDATION env parameter
We added an optional parameter CF_SKIP_SSL_VALIDATION=true for
development environment so that we could some tasks with self
signed certificated.

But now that we use ACM certificates in dev, this is not needed
anymore.

We keep the SKIP_SSL_VALIDATION for the availability tests, as
it was hardcoded as "true" since time ago, and the goal of these
tests is not check that the certificate is valid or not.

We keep it also SKIP_SSL_VALIDATION in the graphite nozzle, as
it is also hardcoded for all environments and it is out of the
scope of this PR.