Revert "Stop sending gorouter logs to syslog"
This reverts commit b3b25e4e9592f96775dc9cd7fbb109d51005f84c.

By enabling blackbox some logs written by the logger command are written to the
disk six times wouch could cause disk fillup and adverse effects on CPU and
disk usage.
Blackbox duplicates the logs handled by the logger command as usually the job
scripts write the logs both to a file and syslog. See [2].

In the latest syslog release notes [1] they only able to partially solve the
issue, that these duplicated logs are not sent to Logsearch but still written
to the disk duplicated.

We decided to not enable blackbox (syslog.forward_files) and wait for a
solution from upstream.

Because of this we have to reenable sending the access logs directly to syslog.

[1] https://github.com/cloudfoundry/syslog-release/releases/tag/v11.1.1
[2] cloudfoundry/syslog-release#31