Run setup-cdn script on concourse
In order to configure cloudfront properly with ACM certificates,
we would need to run the terraform in the concourse worker, that
has the right policies.

In this commit we add a wrapper script that would execute the
step on a adhoc build task in the concourse server, when
passing RUN_ON_CONCOURSE=true.

We use the terraform container and we need to install awscli and jq
to get the IAM cert. This will be removed in the next commit.