Use ACM system cert in cloudfront distributions
We have some cloudfront distributions to be able to use the
system domain with a valid certificate for some applications.

But these cloudfront distributions are using IAM certificates which
we want to replace with ACM certificates.

We add an additional variable to the cloudfront modules to allow
configure an ACM cert arn OR a IAM cert ID. Then update the main
tf file to find the ACM certificate using a datasource and use
it instead of the IAM cert ID.

The ACM certificate must reside in us-east-1[1] to work with
cloudfront, so the data source would use a custom aws provider
pointing to us-east-1 region.

[1] https://docs.aws.amazon.com/acm/latest/userguide/acm-services.html