Unify post-deploy tasks
At the moment, we have two places where some post-deploy tasks are run.
One is sort of inline with the cf-deploy and the other has a dedicated
separate task. We think this will get confusing as to where any future
tasks would come in.

Here, we're moving them back into cf-deploy as it requires less
refactoring to achieve what we need, which is running the tests in the
same time as cf-deploy and post-deploy tasks.