Change ES instance type to compute optimised
This commit changes the instance type used by Elasticsearch to a compute
optimised type as we are currently hitting the CPU limits of the general
purpose instance type.

THe choice of a c4.xlarge was made as it works with the same base AMI we
are using as it does not force enhanced networking which the c5 series
does. Going for the c4.xlarge also means that we keep the RAM around the
same as we are presently using whilst increasing the CPU resources that
we are running short of.

Currently there is an issue filed against the bosh-aws-cpi-release which
states that c5 instances cannot be used currently, this has been
accepted by the bosh team. The issue is located at https://github.com/cloudfoundry-incubator/bosh-aws-cpi-release/issues/77