Add a system domain ELB
We want to be able to serve UAA, the tech docs, and product page websites
from a single ELB.

The configuration for the ELB is based on the existing UAA ELB, because
it has been tuned to avoid UAA downtime during deploys.

We use a wildcard DNS for the following reasons:

* it is a simpler configuration with less Terraform to manage
* it allows for a two-phase deploy where we remove UAA subdomains later
* we want this to be the default route for system domain traffic
* if we change our mind on the previous point any new domains will
  take precedence over the wildcard, as they would be more specific.

We do not remove DNS entries here or take down unnecessary
infrastructure (such as the UAA ELB or the CloudFront distributions) as
they can be done in follow-up deployments to avoid downtime.