Remove Product Pages and Tech Docs DNS entries
We now provide a wildcard DNS entry sending all subdomains of the system
domain to the GoRouters via a dedicated ELB. Having specific DNS entries
for any CloudFoundry routes directly under the system domain is no
longer required.

However, we do have to add a DNS record for the system domain itself
because the CloudFront distributions were redirecting this domain
to the Product Page.

We keep the rest of the CloudFront infrastructure in place to ensure
there is no downtime due to DNS caching. It will be removed in a
subsequent deploy.