Verify CC API is unavailable over HTTP
We recently changed the ELB serving UAA to act as an ELB serving the
product pages and tech docs apps as well as UAA. In order to keep
those apps redirecting from HTTP to HTTPS, we had to allow port 80
traffic on the AWS security group used to enable API traffic.

The security group being used for the system domain ELB is also used
for the separate ELB serving the Cloud Controller API. Therefore the
security group change also changed its behaviour.

We already had a test that HTTP traffic to the Cloud Controller API
timed out without establishing a connection. This test started to fail
because, with the security group allowing port 80, this traffic
reaches the ELB. ELBs seem to listen on port 80 even when configured
otherwise. The connection is established but then closed by the ELB.

Our current hypothesis (by @dcarley) is that there is a transparent
office proxy that accepts (then closes) the connections and this is why
we seem some differences between tests run from the office and those run
from Concourse (within AWS).