Enable continuous smoke tests in dev
We want to run continuous smoke tests in dev, because they can catch
errors caused by things outside of the normal pipeline run, such as
during credential rotation[1]. We use the existing pattern of checking
whether the healthcheck app is deployed to ensure they should be
running.

We use the standard Alpine container to check whether the smoke tests
need to run because this container has no special requirements and the
Docker Hub standard library[2] containers are officially maintained.

We use the govpaas-alerting-dev email address[3] because most people do
not subscribe to this and therefore reduces noise.

[1] https://www.pivotaltracker.com/n/projects/1275640/stories/153435666
[2] https://hub.docker.com/u/library/
[3] https://groups.google.com/a/digital.cabinet-office.gov.uk/forum/#!forum/govpaas-alerting-dev;context-place=myforums