Add monitors for aws elasticache limits
The AWS trusted advisor does not check for the limits of the
elasticache limits for nodes or parameter groups.

We must monitor manually by querying the metrics provided
by the metrics explorer. We have to hardcode the current limits
as AWS does not provide an API to query them.

The current limits are the default ones, 100 nodes and 20 parameter
groups. We set so in the tfvars for each environment.

The monitors would alert on 80% and 90% of the limit, with warning and
critical in each case.