Downgrade stemcell to 3468 series
The `datadog-agent-forwarder` job is failing to start in staging with the
following error:

    Couldn't initialize logging: [Errno 13] Permission denied: '/var/vcap/data/packages/datadog-agent/4ed37e76994196bbc48fbd89ba2d9ffdf9ab7bf3/agent/datadog.conf'

It feels like this could be the same problem we saw with umask changes for
logsearch in 94a70d11. I tried 3541.4, which includes some umask fixes,
but that didn't make any difference.

In order to unblock the pipelines in the meantime we're going to
downgrade all VMs to the latest version of the previous series. We'll
investigate a proper fix for upgrading series later.