Update to use bosh cli v2 commands
We are migrating to the newer BOSH CLI v2, and have updated the docker
image that supplies the bosh command.

The BOSH CLI v2 differs from v1 in two main ways: it is stateless, and
it hyphenates single commands.

For the most part this means that we need to tell bosh which
'environment' and 'deployment' we are operating on with each command or
(as we have decided to do here) use BOSH_CA_CERT, BOSH_ENVIRONMENT and
BOSH_DEPLOYMENT environment variables in each task to preconfigure it.