Ensure we have the bosh-CA.crt resource available
As part of the move to BOSH CLI v2 we need the BOSH_CA_CERT available
throughout the pipeline. The bosh-CA-crt s3 resource is created in
paas-bootstrap but we want to ensure that existing environments
(particulally dev environments) can continue to operate without needing
to execute the paas-bootstrap create-bosh-concourse pipeline.

It should be safe to remove these `extract-bosh-ca-crt` tasks in the
future once the updated create-bosh-concourse pipelines have been
run.