Remove bosh_cli gem from Gemfile
We do not use bosh_cli v1 anymore, so we can remove it from the
Gemfile.

We also remove the separate Gemfile in scripts, as we had
to split them to  work around a bug described in b216ca82

Finally, we install bosh cli v2 in travis, as we need bosh
to run the tests.

Signed-off-by: Andrew Bruce <me@andrewbruce.net>
Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>