Split manifest gen scripts spruce/bosh-interpolate
We want to start using bosh interpolation with variables and ops-files,
rather than using spruce.

In a initial phase, we replace interpolation of variables of any simple
values done with spruce with bosh v2 variables.

The generation of the manifest will be split in two scripts:

 - bosh_interpolate.sh
 - spruce_merge.sh

Note that `bosh interpolate` does not support deep merge when
passing multiple variable files, which causes problems with
variables in sub-trees coming from multiple sources.

To workaround this issue, we rely first on merging the variable
files using spruce in the script bosh_interpolate.sh, but
later we will need to "flatten" the variables key.

Signed-off-by: Andrew Bruce <me@andrewbruce.net>