Flatten variables for secrets and refactor scripts
`bosh interpolate` does not support deep merge of variables from
multiple vars files. Because that, all variables must be flatten
and have a unique root key.

For the case of secrets, we would replace `secrets.` with `secrets_`.

We must refactor the scripts that generate and rotate the secrets:

 - `generate-cf-secrets.rb` would generate flatten keys `secrets_`
 - `generate-cf-secrets.rb` is able to "update" a previous secrets
   YAML file.
 - We refactor how the secrets are rotated, by implement a ruby
   script to filter the secrets to keep, rather than using spruce.

To migrate the existing cf-secrets.yml file in the bucket, we
rely in the script `generate-cf-secrets.rb` that would be called
in the pipeline.

Flatten secrets by replacing secrets_ with secrets.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>