Always populate "old" cert variables
In alphagov/paas-cf#1184 we added logic to rotate the certificates,
which would rename the old certs to append a suffix `_old`

That old certificate would be available in the manifest with the
variables `certs_<cn>_old_key` or `certs_<cn>_old_cert`. But if the
certificate has never been rotated these variables are not defined.

With spruce we could use the operator `||` to give a default value
if the variables were not defined, but that is not possible with
bosh interpolation.

As workaround, we would always generate the variables
`certs_<cn>_old_key` or `certs_<cn>_old_cert` with the current
certificate if the `_old` one does not exist.

This allows us to implement a similar logic as the one implemented
in 7e018f2330117ff678530ed6157b905329fb861c

Signed-off-by: Lee Porte <lee.porte@digital.cabinet-office.gov.uk>