Flatten variables for certificates as certs_
`bosh interpolate` does not support deep merge of variables from
multiple vars files. Because that, all variables must be flatten
and have a unique root key.

In the case of the certificates, we decided rename the suffix to
certs_, so we replace secrets.<var> to certs_<var>

We also remove the generation of ipsec-CA.yml, as it is not being
used.

We generate the YAML file inline in the pipeline, so the script
file_to_yaml.sh is not necessary anymore.

Signed-off-by: Andrew Bruce <me@andrewbruce.net>