Upgrade to logsearch-boshrelease v205.0.1
We want to upgrade to logsearch v205.0.1 [1], that is hosted in
https://github.com/cloudfoundry-community.

Since release v205 logsearch uses ELK stack 5.0.

This release adds ability to disable post-start script for
Elasticsearch jobs, that would prevent failing updates when
BOSH does updates in one AZ at a time.

We also update the parser to connect to the ELB of ES directly.
We found a bug in the parser job in the line [2]: it generates
invalid config when multiple elasticsearch.data_hosts are set,
by writing a string with comma separated values instead of a list
of values. No point to report it as this parser job disappears in
v208.

[1] https://github.com/cloudfoundry-community/logsearch-boshrelease/releases/tag/v205.0.1
[2] https://github.com/cloudfoundry-community/logsearch-boshrelease/blob/v205.0.1/jobs/parser/templates/config/input_and_output.conf.erb#L38