Add jobs to scheduler instance_group
cf-deployment contains the following additional jobs:
 - auctioneer (was brain)
 - cloud_controller_clock (was in api)
 - tps (was in cc_bridge)
 - ssh_proxy (was in access)

As the instance is running more services, we promote the vm_type
to medium.

We remove the cloud_controller_clock from the CC API

We need to add a pair of vm_extensions:
 - ssh_proxy_elb, so the VM is registered for SSH proxy
 - cf_rds_client_sg, so that the cloud_controller_clock can reach
   the CF RDS instance.

The following instance groups would be deleted later to ensure
availability during the change.
 - brain
 - cc_bridge
 - access