Re-add (again) the route_emitter to diego-cell
We delayed this change in the alphagov/pull#1263, and moved to a
dedicated PR.

Repeated commit message:

In cf-deployment they add the route-emitter in the diego cell, but
in local_mode[1] so it only send the routes of the current cell.

We leave the original route_emitter instance_group to avoid downtime
but it should deleted in a future deployment.

NOTE: We assume that is OK to run a global route_emitter and the ones
running in local mode in the cells at the same time.

We put the properties in the specific job to be able to override
the specific option `local_mode`

[1] https://github.com/cloudfoundry/diego-release/blob/develop/jobs/route_emitter/spec#L34-L36