Remove cc_uploader from cc_bridge
It cannot be deployed in two instance_groups at the same time,
otherwise we get this error:

     - Multiple instance groups provide links of type 'cc_uploader'. Cannot decide which one to use for instance group 'api'.
          hector.api.cc_uploader.cc_uploader
          hector.cc_bridge.cc_uploader.cc_uploader

The cc_uploader is required only during pushing applications, so
moving it shall only impact the API availability.

The job is moved to the `api` instance group that is first, so we
expect that the new service is deployed before the old one
gets decommissioned.