Use alpine:latest for all tasks using alpine
In the tasks were we use alpine we use basic `sh` scripts and basic
commands like tar, cat, etc.

There is no need to pin the image to a specific version. We put
the tag to latest to pull the latest. The default is pull latest,
but we want to be explicit to make it clear that  is a informed
decision.