Enlarge the parser VMs in staging and prod.
They are not presently keeping up with the volume of incoming messages.
This is partly due to issues with the logging stack not being balanced,
meaning that often one of the queues is empty with all incoming messages
going to the other queue.

We believe that the underlying cause for this unbalance is down to how
the ingestors are load-ballanced. Because all the VMs use persistent TCP
connections to the ingestors via the ELB, when one of the ingestors is
restarted this results in all the VMs connecting to the other one, and
therefore all log messages ending up in one queue.

A longer term fix would be to restructure things so that we use a single
shared queue (possibly using AWS Elasticache), however given we hope to
have migrated off this stack before long enklarging the parsers is the
simplest thing we can do to ensure things don't back up.