Revert "Add the route_emitter to diego-cell"
This reverts commit 80eb427f3b3f8a25950281cf7351aff1762a28ce.

We decided to delay this change to a specific separated PR, as
it is an important change in the architecture and there is some
risk when running route_emitter in global and local_mode.

We will do it later in a dedicated PR with only that change
that would be easy and quick to revert in case of issues.