Move specific consul and metron properties to job
As part of migrating to cf-deployment, move all the global
properties for consul to the specific job.

As consul agent and metron are added to all the instance
groups in the file we use YAML anchors to not repeat ourselves.

In the case of the consul servers, we need to copy all the
properties again to simply set `consul.agent.mode: server`
because YAML `<<:` operator does not do a deep merge.

For the sake to keep the minimum changes as possible, in
some cases we keep some unnecessary config like the
`consume` for some jobs, or override the log_level.

Here we remove a spurious `max_in_flight: 1` which was added
to the `properties:`[1] of the job instead of `update:`. It
likely never worked.

[1] https://github.com/alphagov/paas-cf/commit/b69a10e5b3dc5154a51ac4e570680c3e26e9c6fd