Pull manifest properties from new location
All the properties in the manifest are now defined in each
specific job in each specific instance_group, following the
convention done in cf-deployment.

But we pull properties from the generated manifest
in different places in the pipeline, and other scripts.

We must update them to pull the properties from the specific
jobs in the instance profiles.