Adapt pipeline and scripts to new property layout
As we migrate to the cf-deployment, now we have all properties
inside of specific jobs in the instance groups, rather than global
properties.

But we extract several values from the cf-manifest in the deployment
pipeline and some scripts. We must adapt them to consume the
properties from the right location.

In some cases, like the CC api url, we have dropped the property
with that value. We generate the value from variables in the
pipeline itself.