Adapt manifest unit tests to new property layout
As we migrate to the cf-deployment, now we have all properties
inside of specific jobs in the instance groups, rather than global
properties.

We refactor the unit tests to check the values in the new locations.

To simplify the code, we copy the utility class PropertyTree
from val_from_yaml.rb, which allows retrieve a key using
the syntax 'key1.key2.key2'