Split test config from manifest
We were adding the smoke-tests and acceptance-tests into
the manifest to later pull it from the rendered manifest
to pass it as a .json file.

But there is no need for this, we can simply use bosh
interpolation to populate the variables and convert the
result to json without using the rendered manifest.

This way we do not add any unnecessary properties to the final
manifest.

In the case of the smoke test, we need to pass `user` and `password`
instead of `admin_user` and `admin_password`.