Temporarily increase threshold for Compose memory
This is constantly warning at the moment. We're attempting to increase
the size of our Compose cluster, but we don't know how long it will take
to go through procurement.

Increase the critical threshold so that it still tells us when we've
reached a level that makes it unlikely we'll be able to provision any
more service instances. This is based behaviour that we observed last
week. Remove the warning threshold because we no longer have enough
headroom for this to be useful.

This commit should be reverted when the cluster has been resized.