Use a constant name for the UAA JWT signing keys
Previously we used `default` and `previous` as key names so we can support
rotation, but the issued tokens store the key id (kid) they have been created
with and a token will be always verified with that key id. See [1].

When we rotated the keys we generated a new key as `default` and all existing
tokens were verified with the new key and not the previous one.

We're going to use the SHA1 hash of the keys instead of `default` and
`previous` so we can fully support rotation.

[1] https://github.com/cloudfoundry/uaa/blob/master/docs/UAA-APIs.rst#get-the-token-signing-key-get-token-key