Use bosh-job instead of job in datadog queries
The metrics from the nodes in datadog would have two different
tags to refer to the `instance_group` (formerly `job`): `job:*`
and `bosh-job:*`.

The first one `job:*` is added via AWS, as the VMs are created in
AWS with the same tags. The tags are added during VM creating by
the CPI.

The second `bosh-job:*` would be provided for each datadog agent,
that gets updated whenever the instance is updated by bosh.

Because this, it is more convenient and reliable to use `bosh-job:*`
as it will be updated even when we do a `instance_group` name
migration (e.g. `cell` => `diego_cell`)

But note that the metrics from the firehose nozzle (e.g. `cf.*`)
would instead be tagged with `job:`.