Use BOSH links for Consul agents
We want to be closer to the upstream set up [1] so we are using BOSH links [2] instead of setting the configuration for every consul_agent job.
The parameters we are setting for the Consul server/clients are provided by the links [3].

We also updated all consul_agent's properties to match the upstream configuration.

Differences from upstream:
 - we set consul.agent.log_level to null

The consul.agent.dns_config params were removed as those are the default values.

[1] https://github.com/cloudfoundry/cf-deployment/blob/v1.14.0/cf-deployment.yml
[2] https://bosh.io/docs/links.html
[3] https://github.com/cloudfoundry-incubator/consul-release/blob/v191/jobs/consul_agent/spec#L22-L47