Change certs vars to look like cf-deployment
We are converting our manifest to look more like cf-deployment.

In this commit we will define all the internal certificates as
they are defined in cf-deployment, and add them as variables
in the manifest [1]. This will enable the capability of
generating certificates with `bosh interpolate`[1]

All the certificates will be renamed to match the name and convention
as it is in cf-deployment:
 - certs_<name>_key => <name>.private_key
 - certs_<name>_cert => <name>.certificate

In the definition of the certificates as variables, we add dedicated
CA certificates for each service and mutual TLS connection, as it
is done in cf-deployment. Before we were using bosh-CA for all
the certificates.

We can still initialise these certificates with the old values as
we do in migration job that initialises the bosh var store.

We add uaa_jwt_signing_key_id to support the UAA rotation. Previously
was a hash of the private key, but after will be replaced
by a random string (i.e. a password)[2]

We change the properties: diego.rep.ca_cert,
diego.rep.server_cert,  diego.rep.server_key, which are deprecated
and have been moved to the corresponding tls.*, so they should be
the same. In the migration script we decided to use the `rep_server`
in both, as it contains the domain name.

[1] https://bosh.io/docs/variable-types.html
[2] https://github.com/alphagov/paas-cf/pull/1272/commits/b16e486d9bd65ed68e8b054f29ef4f12bcbc30b0