Make API properties look more like upstream
The ordering of properties has been made more consistent with upstream
cf-deployment.yml, so makes it easier compare the two by eye when
reading source code.

Also, we remove a lot of properties that were being set to their default
values. Previously we justified setting them explicitly for readability,
but we now place a higher value on having a manifest that looks like the
upstream manifest.