Switch Cloud Controller DNS to system domain ELB
Now that Cloud Controller registers its route with GoRouter we can
start directing traffic to the system domain ELB. The redundant
infrastructure can be removed in a future commit to avoid downtime.

We don't need to actually CNAME to the ELB domain name because there
is a wildcard DNS record for the system domain in place, so the record
can just be removed.