Check var store certificates expire date
We use bosh to generate the new certificates, and it creates
them with a life of 1y. We need to monitor if any certificate
is about to expire.

For that we add a new task in health that would run everyday
and check if any certificate is about to expire. Currently we
alert if the expire date is < 30 days.