Revert "Merge pull request #1287 from alphagov/move-cc-behind-gorouter-part2"
This reverts commit a38114dfaf4071e7868507a33037c70adb40e05d, reversing
changes made to f9d3e0350de98ef295882b5037d0fe9f8c7a17c2.

We switched the CF API endpoint over to the cf_router_system_domain ELB but we overlooked the idle timeout which is much lower (19s) than it was on the API ELB (300s). The lower idle timeout caused `cf push` commands taking longer than 19s to be aborted.

We'll revert this change for now so we'll have time to figure out the right solution.