Increase idle timeout for the system domain ELB
We are moving the Cloud Controller API behind the Gorouter and behind
the system domain ELB. Deploying this the first time, we discovered that
the system domain ELB had a 19-second timeout. This caused long-running
CC API tasks (e.g., deploying app) to fail.

We increase to the same value as the CloudController's current ELB.
This may need changing, given traffic now goes via GoRouter, but this
will be a good starting point.

The connection_draining_timeout has been increased to 110 seconds to
fill the time allowed via GoRouter's drain_wait configuration . When
GoRouter starts draining it will start failing the healthchecks. The 10
seconds it takes to register as unhealthy with the ELB, plus the 110
seconds allowed for connection draining, matches the 120 total
drain_wait time configured on the GoRouters.