Generate cf-manifest-pre-vars for cert rotation
The `variables` block is pruned from `cf-manifest.yml` after all of the
missing variables have been created/updated from `cf-vars-store.yml`.

We need the `variables` for certificate rotation, so that we can
identify whether a cert is a CA or not, so we need to generate a
separate copy of the manifest without the variable store interpolation.