Vertically scale queue VMs
This increases the queue VMs from t2.small to m3.medium instances.
We have been seeing the queue VMs maxing out on CPU and running out
of credits. The m3.medium instances offer twice the vCPU and a fixed
performance.

This has been made into an environment-specific configuration because
we anticipate our logging stack to be scaled in numerous ways as we
grow, so there will be large cumulative cost savings.