Upgrade cf components to match cf-deployment 1.21
To address CVE-2018-1266 we needed to upgrade to CAPI 1.52.0, which
required bumping Diego to 2.1.0. We duplicated the release versions
advertised in the cf-deployment 1.21 release as a known compatibility
base.

The change to Diego 2.1.0 involved a move to all components using the
loggregator v2 API (enabled by setting `loggregator` properties on each
diego component (rep/auctioner/etc) in the manifest. Without configuring
the components to use the new v2 API we observed smoke test failures due
to application logs not being received through `cf logs`.

The change to Diego 2.1.0 enables the locket server by default for all
components. We did not want to enable this right now as our focus was on
getting out the CVE fix so disabled it via the locks.locket.enabled
properties.  We should expect to enable it soon as there is an
expectation that cf-deployment will be using it for all components.