Use new version of cf-cli in CATs tests
The new cf-acceptance-tests blessed by cf-deployment v1.21.0
use a new command: cf add-network-policy, that was added in
cf 6.35.0, so the tests fail.

We upgrade to the latest cf-cli version used in the docker container
that runs the cf-acceptance-tests in the PR
paas-docker-cloudfoundry-tools#123, and in this commit we point
to the tagged image.