Ignore the test for network policies in v3 tasks
In the latest blessed cf-acceptance-tests for cf-deployment v1.21,
they added a test for network policies when running v3 tasks.

Network policies is a feature from the private networking, which
we still do not have enabled in our platform, so the test fail.

We need to disable this test, by adding a new regex to the SKIP_REGEX
expression passed to ginkgo.

In this commit we also refactor how this expression is built,
by splitting each matching string in a new line with spaces. This
way it is easier to read, add new entries, search, and track in the
git history.