Enable back v3 API tests
In the commit 80494a60eb55c86db50d7d59133f6edc76354b6a we disabled
the v3 api tests because they were failing. Back then they were
a experimental feature, and we were not using it.

But these endpoints are starting to be officially supported upstream
and some tenants are starting to use them.