Periodically dump infrastructure state to bucket
We're building a (somewhat) live infrastructure map dashboard to get an
overview of deployed components.

We create a job that can be run in the pipeline to create a JSON
snapshot of the state of what the BOSH deployment looks like (via bosh
director API) and a dump of EC2/ELB info which gets written to the state
bucket.

To allow access to the generated file a presigned URL for the state is
created and writen to a secrets file to be consumed by an application.