Stop using bosh-CA if not needed
Since the beginning of the project we have been using the bosh CA,
created in paas-bootstrap[1], as the main CA for all the certificates.

We used it also for the system endpoints (API, uaa, etc) and the
application endpoints in development environments.

Because that, we had to import this bosh CA as a valid CA for
every task which interacted with the API, UAA or Bosh.

But in recent stories we changed this:

 - API and UAA use ACM certificates from AWS, that are valid with
   the default CA installed in all the systems (not self signed) [2]
 - We use dedicated CAs for all the internal certificates[3]

Now we do not need to pull the bosh CA in any tasks except the
ones that interact with the BOSH API.

We remove all these references. We also remove the tasks that
were uncompressing the bosh-CA.crt from the bosh-CA.tar.gz file.
This is only needed in paas-bootstrap[4].

[1] https://github.com/alphagov/paas-bootstrap/blob/40d44faf4aa43a061755e3e8f00017df81f3559b/concourse/pipelines/create-bosh-concourse.yml#L396-L434
[2] https://github.com/alphagov/paas-cf/pull/1202
[3] https://github.com/alphagov/paas-cf/pull/1292
[4] https://github.com/alphagov/paas-bootstrap/blob/40d44faf4aa43a061755e3e8f00017df81f3559b/concourse/pipelines/create-bosh-concourse.yml#L431-L434