Add manifest unit test for certificate rotation
In the PR alphagov/paas-cf#1292 we implemented internal certificate
rotation by appending a `_old` certificate for each CA cert to rotate.

When rotating the certs we will:

 1. Deploy the new and old certificate for every CA.
 2. Regenerate and deploy the leaf certs
 3. Redeploy deleting the old certificate.

To support this, we must append both the new and old certificate
everywhere a CA certificate is used. We do so by using
the opsfile `cf-manifest/manifest/operations/010-cert-rotation.yml`.

This requirement is easy to omit during upgrades and changes.
Implement a end to end integration test of the certificate rotation
is complicated and slow, so we want to add some unit test of the
feature.

In this commit we add a test that would check the manifest to:

 1. Find every CA cert variable with a corresponding `_old` variable
 2. Check that every time the variable is used, it has a corresponding
    appended `_old.certificate`