Split continuous smoke test assert and payload
During alphagov/paas-cf#1232 we added a task that would check
if the healthcheck is deployed and if the environment is a dev
environment, skipping the execution if it is not with a `try`
block.

But that semantically makes the job to always be "green" because
all the sequence is under a `try` block, even when the smoke test
actually fails. Although the email alert is sent, this prevents
our datadog monitor to be trigger because it checks if there
was a "green" build in the last X minutes.

We can rewrite the flow of the job so that the assertion runs first,
and if it fails, it stops all the execution of the task. The email
would be sent only if the smoke test fails.

In development we do not have datadog monitoring, so it is
fine to have this job as failed.